requirements.txt
#db.py
sqlmodel~=0.0.24


#main.py
fastapi~=0.116.1
uvicorn~=0.35.0

# Code-quality tools
pre-commit>=3.7
black>=24.4
isort>=5.13
ruff>=0.4
flake8>=7.0

# user/auth
python-dotenv~=1.0
python-jose[cryptography]~=3.3
passlib[bcrypt]~=1.7
python-multipart

# test
pytest~=8.4.1
pytest-asyncio



# httpx
h11~=0.16.0
pip~=25.1.1
typing_extensions~=4.14.1
sniffio~=1.3.1
pyflakes~=3.4.0
pycodestyle~=2.14.0
starlette~=0.47.1
pydantic~=2.11.7
pydantic_core~=2.33.2
anyio~=4.9.0
click~=8.2.1
greenlet~=3.2.3
annotated-types~=0.7.0
SQLAlchemy~=2.0.41
httpx~=0.28.1